All_requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
numpy==1.24.4
packaging==23.1
pandas==2.0.3
protobuf==4.21.12
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.6
zipp==3.16.2
